[Format] change files to correct locations
