+++ srcs/main.c
@@ -16,25 +16,6 @@ int can_mino_move(t_mino *check_mino, char **table)
 	return true;
 }
 
-void	run_tetris_game(t_player *player)
-{
-	/* init_struct */
-	player = init_struct();
-
-	/* init_game */
-	init_game();
-
-	generate_mino(player->mino);
-	if(!can_mino_move(player->mino, player->table->table_array))
-		player->table->is_game_on = false;
-	while (player->table->is_game_on)
-	{
-		print_output_according_to_fps(player);
-		key_events(player);
-		time_elapse_event(player);
-	}
-}
-
 int	main(void)
 {
 	t_player	*player;
