+++ srcs/struct.c
@@ -1,19 +1,5 @@
 #include "tetris.h"
 
-char	**init_table(void)
-{
-	char	**table;
-
-	table = (char **)xmalloc(sizeof(char *) * (ROWS + 1));
-	for (int i = 0; i < ROWS; i++)
-		table[i] = (char *)xmalloc(sizeof(char) * (COLUMNS + 1));
-	table[ROWS] = NULL;
-	for (int i = 0; i < ROWS; i++)
-		for (int j = 0; j < COLUMNS; j++)
-			table[i][j] = 0;
-	return (table);
-}
-
 t_player	*init_struct(void)
 {
 	t_player	*player;
@@ -30,4 +16,26 @@ t_player	*init_struct(void)
 	player->table->score = 0;
 	player->table->is_game_on = true;
 	return (player);
+}
+
+void	destruct_table_struct(t_table *table)
+{
+	free_array(table->table_array);
+	free(table);
+	table = NULL;
+}
+
+void	destruct_mino_struct(t_mino *mino)
+{
+	free_array(mino->mino_array);
+	free(mino);
+	mino = NULL;
+}
+
+void	destruct_player_struct(t_player *player)
+{
+	destruct_table_struct(player->table);
+	destruct_mino_struct(player->mino);
+	free(player);
+	player = NULL;
 }
