+++ srcs/table_alloc.c
@@ -1,0 +1,15 @@
+# include "tetris.h"
+
+char	**init_table(void)
+{
+	char	**table;
+
+	table = (char **)xmalloc(sizeof(char *) * (ROWS + 1));
+	for (int i = 0; i < ROWS; i++)
+		table[i] = (char *)xmalloc(sizeof(char) * (COLUMNS + 1));
+	table[ROWS] = NULL;
+	for (int i = 0; i < ROWS; i++)
+		for (int j = 0; j < COLUMNS; j++)
+			table[i][j] = 0;
+	return (table);
+}
